requirements.txt

bcrypt>=5.0.0
eventlet>=0.40.3
flask>=3.1.2
flask-cors>=6.0.1
flask-jwt-extended>=4.7.1
flask-socketio>=5.5.1
google-generativeai>=0.3.0
gunicorn>=21.2.0
openai>=2.0.0
pillow>=11.3.0
psycopg2-binary>=2.9.10
pymongo>=4.15.1
python-dotenv>=1.1.1
requests>=2.32.5
sqlalchemy>=2.0.43
stripe>=13.0.0
